+++ project/Cargo.toml
@@ -4,17 +4,3 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-crossbeam-channel = "0.5.13"
-toml = "0.8.19"
-wg_2024 = { git = "https://github.com/WGL-2024/WGL_repo_2024.git", features = ["serialize"] }
-game_of_drones = { git = "https://github.com/AP-Game-Of-Drones/GoD_Library.git" }
-bagel_bomber = { git = "https://github.com/daw-dev/bagel-bomber.git" }
-drone_bettercalldrone = { git = "https://github.com/BetterCallDrone/drone_BetterCallDrone.git" }
-getdroned = { git  = "https://github.com/GetDroned/Drone.git" }
-d-r-o-n-e_drone = { git = "https://github.com/AP-2024-25-D-R-O-N-E/drone.git" }
-null-pointer-drone = { git = "https://github.com/The-Null-Pointer-Patrol/null-pointer-drone.git" }
-ap2024_unitn_cppenjoyers_drone = { git  = "https://github.com/Cpp-enjoyers/drone.git" }
-rustafarian-drone = { git = "https://github.com/Rustafarian-Unitn/rustafarian-drone.git" }
-rust-roveri = { git = "https://github.com/RustRoveri/rust-roveri.git" }
-rusteze_drone = { git = "https://github.com/Rusteze-AP/drone.git" }
-dr_ones = { git = "https://github.com/Dr-Ones/drone" }
