+++ project/client/Cargo.toml
@@ -9,3 +9,5 @@ crossbeam-channel = "0.5.13"
 controller = { path = "../controller" }
 bevy = "0.11"
 image = "0.24"
+topology = { path = "../topology" }
+fragmentation_handling = { path = "../fragmentation_handling" }
