+++ project/fragmentation_handling/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "fragmentation_handling"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+wg_2024 = { git = "https://github.com/WGL-2024/WGL_repo_2024.git", features = ["serialize"] }
+bevy = "0.11"
+image = "0.24"
