+++ project/src/main.rs
@@ -1,1 +1,3 @@
-fn main(){}
+fn main() {
+    println!("Hello, world!");
+}
