+++ project/topology/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "topology"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+wg_2024 = { git = "https://github.com/WGL-2024/WGL_repo_2024.git", features = ["serialize"] }
+rand = "0.8.5"
